src/translator/requirements.txt
yt-dlp
openai
python-dotenv
pydub
groq
pytest 
pytest-mock
vimeo_downloader
langsmith
langchain
pika
azure-storage-blob
validators
pydantic